backend/requirements.txt

fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
easyocr>=1.7.0
pillow>=9.5.0
numpy>=1.24.0
pydantic>=1.10.0
aiofiles>=23.1.0
